Feat: 게시판 좋아요, 조회수 캐싱 및 푸터 하단 고정, 일부 UI수정

+++ backend/src/main/java/com/ssafy/enjoytrip/features/tripost/application/dto/TripostListItemDto.java
@@ -20,9 +20,9 @@ public class TripostListItemDto {
     private String uid;
     private String name;
     private LocalDateTime createdAt;
-    private Integer commentCount;
-    private Integer viewCount;
-    private Integer likeCount;
+    private long commentCount;
+    private long viewCount;
+    private long likeCount;
     private List<WaypointSnapshotDto> waypoints;
     private List<WaypointImageDto> images;
 }
